changed purchase structure and changed type that is related with money

+++ src/main/java/com/jun/smartlineup/payment/dto/PaymentPayRequestDto.java
@@ -3,15 +3,17 @@ package com.jun.smartlineup.payment.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PaymentPayRequestDto {
     private String customerKey;
-    private Long amount;
+    private BigDecimal amount;
     private String orderId;
     private String orderName;
     private Integer cardInstallmentPlan = 1;
     private String customerEmail;
     private String customerName;
-    private Long taxFreeAmount;
+    private BigDecimal taxFreeAmount;
 }
